add new test and new method
